move all dash result writing APIs to a new dedicated file
