+++ orchagent/Makefile.am
@@ -132,6 +132,7 @@ orchagent_SOURCES = \
             dash/dashhaorch.cpp \
             dash/dashhafloworch.cpp \
             dash/dashportmaporch.cpp \
+            dash/dashresulthelper.cpp \
             twamporch.cpp \
             stporch.cpp \
             nexthopkey.cpp \
